kcms/lookandfeel: Add auto dark-light lnf kded module
This change introduces a kded module that synchronizes the look-and-feel
package to the dark-light cycle.

The dark look-and-feel package will be applied during night, and the light
look-and-feel package will be applied during daylight.

In order to minimize interruptions, the global theme will be changed
only if the computer has been idle for a certain amount of time (currently,
it is 5 seconds).

The module can be enabled by running

    kwriteconfig6 --notify --file kdeglobals --group KDE --key AutomaticLookAndFeel --type bool true